database/requirements.txt
pinecone>=7.3.0
python-dotenv>=1.1.1
pypdf>=5.1.0
